+++ sciml_bench/benchmarks/slstr_cloud/slstr_cloud.py
@@ -1,12 +1,59 @@
-from sciml_bench.core.runtime import RuntimeIn, RuntimeOut
-from sciml_bench.core.utils import SafeDict
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# slstr_cloud.py
 
+# SciML-Bench
+# Copyright © 2021 Scientific Machine Learning Research Group
+# Scientific Computing Department, Rutherford Appleton Laboratory
+# Science and Technology Facilities Council, UK.
+# All rights reserved.
+
+from sciml_bench.core.utils import MultiLevelLogger
 import yaml
 import tensorflow as tf
 import horovod.tensorflow as hvd
 
+from sciml_bench.core.runtime import RuntimeIn, RuntimeOut
+
 from sciml_bench.benchmarks.slstr_cloud.train import train_model
-from sciml_bench.benchmarks.slstr_cloud.inference import inference
+from sciml_bench.benchmarks.slstr_cloud.data_loader import load_datasets
+from sciml_bench.benchmarks.slstr_cloud.model import unet
+
+
+# from sciml_bench.benchmarks.slstr_cloud.inference import inference
+
+
+# Sets the target device
+def set_target_devices(use_gpu: bool, log: MultiLevelLogger) -> bool:
+    if use_gpu:
+        try:
+            tf.config.set_visible_devices([], 'GPU')
+            gpus = tf.config.experimental.list_physical_devices('GPU')
+            for gpu in gpus:
+                tf.config.experimental.set_memory_growth(gpu, True)
+            if gpus:
+                tf.config.experimental.set_visible_devices(
+                    gpus[hvd.local_rank()], 'GPU')
+
+            if hvd.rank() == 0:
+                log.message('Using GPU(s)')
+                log.message(f"Num GPUS: {len(gpus)}")
+                log.message(f"Num ranks: {hvd.size()}")
+            return True
+        except:
+            use_gpu = False
+
+
+    if not use_gpu:
+        try:
+            tf.config.set_visible_devices([], 'CPU')
+            log.message('Using CPU(s)')
+            return True
+        except:
+            log('Failed to use CPUs! Terminating.')
+            return False
+
 
 #####################################################################
 # Training mode                                                     #
@@ -16,101 +63,87 @@ from sciml_bench.benchmarks.slstr_cloud.inference import inference
 def sciml_bench_training(params_in: RuntimeIn, params_out: RuntimeOut):
     """
     Main entry for `sciml_bench run` in training mode
-
-    :param params_in: runtime input of `sciml_bench run`, useful components:
-        * params_in.start_time: start time of running as UTC-datetime
-        * params_in.dataset_dir: dataset directory
-        * params_in.output_dir: output directory
-        * params_in.bench_args: benchmark-specific arguments
-    :param params_out: runtime output of `sciml_bench run`, useful components:
-        * params_out.log.console: multi-level logger on root (rank=0)
-        * params_out.log.host: multi-level logger on host (local_rank=0)
-        * params_out.log.device: multi-level logger on device (rank=any)
-        * params_out.system: a set of system monitors
     """
-    pass
 
-    # # -----------------------------------
-    # # initialize horovod and smlb_monitor
-    # # -----------------------------------
-    # # initialize horovod
-    # hvd.init()
+    # initialize horovod
+    hvd.init()
 
-    # # initialize smlb_monitor with hvd.rank() and hvd.local_rank()
-    # # In this example, we will log the whole process on console and the
-    # # epoch loop on device, leaving the host logger unactivated.
-    # params_out.activate(rank=hvd.rank(), local_rank=hvd.local_rank(),
-    #                       activate_log_on_host=False,
-    #                       activate_log_on_device=True, console_on_screen=True)
+    # initialize monitor with hvd.rank() and hvd.local_rank()
+    params_out.activate(rank=hvd.rank(), local_rank=hvd.local_rank(),
+                          activate_log_on_host=False,
+                          activate_log_on_device=True, console_on_screen=True)
 
-    # # in this example, we will log sub-processes in console while
-    # # using the device logger only in epoch loop for clarity
-    # console = params_out.log.console
+    console = params_out.log.console
 
-    # # top-level process
-    # console.begin('Running benchmark slstr_cloud in training mode.')
-    # console.message(f'hvd.rank()={hvd.rank()}, hvd.size()={hvd.size()}')
+    # top-level process
+    console.begin('Running benchmark slstr_cloud in training mode.')
+    console.message(f'hvd.rank()={hvd.rank()}, hvd.size()={hvd.size()}')
 
-    # # ------------------------------------------
-    # # input arguments and tensorflow environment
-    # # ------------------------------------------
-    # console.begin('Handling input arguments and tensorflow environment')
-    # # default arguments
-    # default_args = {
-    #     # tensorflow env
-    #     'seed': 1234,
-    #     'use_cuda': True,
-    #     # hyperparameters
-    #     'learning_rate': 0.001,
-    #     'epochs': 30,
-    #     'batch_size': 32,
-    #     'wbce': .5,
-    #     'clip_offset': 15,
-    #     'train_split': .8,
-    #     'crop_size': 80,
-    #     'no_cache': False,
-    #     # workflow control
-    #     'load_weights_file': '',  # do training if this is empty
-    # }
-    # args = params_in.bench_args.try_get_dict(default_args)
 
-    # train_data_dir = params_in.dataset_dir / 'one-day'
-    # test_data_dir =  params_in.dataset_dir / 'ssts'
+    default_args = {
+        # tensorflow env
+        'seed': 1234,
+        'use_gpu': True,
+        # hyperparameters
+        'learning_rate': 0.001,
+        'epochs': 30,
+        'batch_size': 32,
+        'wbce': .5,
+        'clip_offset': 15,
+        'train_split': .8,
+        'crop_size': 80,
+        'no_cache': False,
+    }
 
-    # # tensorflow environment
-    # tf.random.set_seed(args['seed'])
-    # console.message(f'Random seed: {args["seed"]}')
-    # if args['use_cuda']:
-    #     try:
-    #         # Disable all GPUS
-    #         tf.config.set_visible_devices([], 'GPU')
-    #         visible_devices = tf.config.get_visible_devices()
-    #         for device in visible_devices:
-    #             assert device.device_type != 'GPU'
-    #     except:
-    #         # Invalid device or cannot modify virtual devices once initialized.
-    #         pass
 
-    # # save actually used arguments
-    # if hvd.rank() == 0:
-    #     args_file =  params_in.output_dir / 'arguments_used.yml'
-    #     with open(args_file, 'w') as handle:
-    #         yaml.dump(args, handle)
-    #     console.message(f'Arguments used are saved to:\n{args_file}')
+    with console.subproc('Parsing input arguments'):
+        args = params_in.bench_args.try_get_dict(default_args)
 
-    # console.ended('Handling input arguments and tensorflow environment')
 
-    # # Train model or load weights from file
-    # if args['load_weights_file'] == '':
-    #     console.begin('Training model')
-    #     train_model(train_data_dir, args, params_in, params_out)
-    #     args['load_weights_file'] =  params_in.output_dir / 'model.h5'
-    #     console.ended('Training model')
+    tf.random.set_seed(args['seed'])
+    console.message(f'Random seed: {args["seed"]}')
 
-    # # Inference
-    # console.begin('Inference')
-    # inference(args['load_weights_file'], test_data_dir, args, params_in, params_out)
-    # console.ended('Inference')
+    train_data_dir = params_in.dataset_dir / 'one-day'
+    test_data_dir =  params_in.dataset_dir / 'ssts'
+
+
+    if set_target_devices(args['use_gpu']) == False:
+        console.ended('Running benchmark slstr_cloud in training mode.')
+        return 
+
+            
+    # save arguments
+    if hvd.rank() == 0:
+        args_file = params_in.output_dir / 'training_arguments_used.yml'
+        with console.subproc('Saving arguments to a file'):
+            with open(args_file, 'w') as handle:
+                yaml.dump(args, handle)
+
+    # load the datasets
+    with console.subproc("Loading datasets"):
+        train_dataset, test_dataset  = load_datasets(dataset_dir=train_data_dir, args=args)
+    
+    # build the UNet model
+    with console.subproc('Creating the model'):
+        model = unet(input_shape=(256, 256, 9))
+        
+        
+    # train model
+    with console.subproc("Training the model"):
+        history = train_model(train_dataset, test_dataset, model, args, params_in, params_out)
+
+
+    # save history
+    with console.subproc('Saving training history'):
+        history_file = params_in.output_dir / 'training_history.yml'
+        with history_file.open('w') as handle:
+            yaml.dump(history, handle)
+
+
+    console.begin('Running benchmark slstr_cloud in training mode.')
+
+
+
 
 
 
@@ -120,17 +153,12 @@ def sciml_bench_training(params_in: RuntimeIn, params_out: RuntimeOut):
 def sciml_bench_inference(params_in: RuntimeIn, params_out: RuntimeOut):
     """
     Main entry for `sciml_bench run` in inference mode.
-
-    :param params_in: runtime input of `sciml_bench run`, useful components:
-        * params_in.start_time: start time of running as UTC-datetime
-        * params_in.dataset_dir: dataset directory
-        * params_in.output_dir: output directory
-        * params_in.bench_args: benchmark-specific arguments
-    :param params_out: runtime output of `sciml_bench run`, useful components:
-        * params_out.log.console: multi-level logger on root (rank=0)
-        * params_out.log.host: multi-level logger on host (local_rank=0)
-        * params_out.log.device: multi-level logger on device (rank=any)
-        * params_out.system: a set of system monitors
     """
 
     pass
+
+
+    # Inference
+    # console.begin('Inference')
+    # inference(args['load_weights_file'], test_data_dir, args, params_in, params_out)
+    # console.ended('Inference')